requirements.txt
# Core
langchain
langchain-chroma
langchain-community
langchain-openai
langchain-ollama
chromadb
rank-bm25
python-dotenv
pymupdf
uvicorn
fastapi
numpy
streamlit
transformers==4.40.2
sentence-transformers==2.7.0

# Local LLM and OCR
ollama
pytesseract
Pillow

# Torch (installing from pip works on many setups; if it fails, see notes below)
torch

paddleocr
paddlepaddle